# Conflicts:
#	src/main/resources/application-dev.properties
#	target/classes/application-dev.properties
#	target/classes/com/spring/jwt/Car/CarServiceImpl.class
#	target/classes/com/spring/jwt/dealer/DealerServiceImpl.class
#	target/classes/com/spring/jwt/dto/DealerDTO.class
#	target/classes/com/spring/jwt/dto/UserDTO.class
#	target/classes/com/spring/jwt/entity/Dealer.class
#	target/classes/com/spring/jwt/entity/PendingBooking.class
#	target/classes/com/spring/jwt/service/impl/UserServiceImpl.class

+++ src/main/java/com/spring/jwt/service/impl/UserServiceImpl.java
@@ -377,7 +377,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserDTO getUserById(Long id) {
+    public UserDTO getUserById(Integer id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundExceptions("User not found with id: " + id));
         
@@ -415,7 +415,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserDTO updateUser(Long id, UserUpdateRequest request) {
+    public UserDTO updateUser(Integer id, UserUpdateRequest request) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundExceptions("User not found with id: " + id));
 
@@ -437,7 +437,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserProfileDTO getUserProfileById(Long id) {
+    public UserProfileDTO getUserProfileById(Integer id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundExceptions("User not found with id: " + id));
